fix: rename utils to nodes
